renamed netaccept.go to server.go
